extracted interfaces into separate file
