+++ kanban-fire/src/app/feature-swim-lanes/task-dialog/task-dialog.ts
@@ -1,0 +1,11 @@
+import { Task } from "../task/task";
+
+export interface TaskDialogData {
+    task: Partial<Task>;
+    enableDelete: boolean;
+}
+
+export interface TaskDialogResult {
+    task: Task;
+    delete?: boolean;
+}
